Better hashing error handling in checksum file creation modes, hashing task reporting in these modes too
When an error happens during hashing of a file, it is now
recorded in the checksum file itself along with a 0 checksum to
make it extremely unlikely that the error will be ignored when
later using the file to hash. Fixes #6.

Like normal hashing modes, stats about the hashing task will be
reported after hashing in checksum file creation modes too.